Don't pull ganja.js when this module is installed
As previously written, it makes it very difficult to roll back ganja.js.
After this change, the user will get the version of ganja.js that pyganja was tested with, not just blindly the latest version.
This is consistent with what they will get with PyPI

There's nothing to stop a user pulling a newer version locally after this change, if they still want the latest.